refactor: streamline image generation process and enhance data transformation utilities
